fix: 자료형 버그 수정 및 AI 분석 기능 테스트 추가

+++ src/main/java/org/example/animation/dto/AnimationDetailResponse.java
@@ -9,7 +9,7 @@ public record AnimationDetailResponse(
         String jsonData,
         Long languageId,
         String languageName,
-        Long creatorUserNumber,
+        Integer creatorUserNumber,
         String creatorUsername,
         LocalDate createdAt
 ) {
